SubsidenceChron

Documentation for SubsidenceChron.jl.

SubsidenceChron.SubsidenceStratMetropolis

SubsidenceChron.find_formation_depths

SubsidenceChron.find_formation_heights

SubsidenceChron.subsidenceparams

SubsidenceChron.SubsidenceStratMetropolis — Method

SubsidenceStratMetropolis(smpl::ChronAgeData, config::StratAgeModelConfiguration, therm::ThermalSubsidenceParameters, subsidence_strat_depths, Sμ, Sσ, beta_ip, t0_ip;
    subsidencebotom=minimum(smpl.Height),
    subsidencetop=maximum(smpl.Height),
    lithosphere = Normal(125000, 100),
)

Runs the main SubsidenceChron.jl age-depth model routine given a set of Gaussian age constraints specified in the smpl struct, an age-depth model configuration specified in the config struct, thermal subsidence parameters defined in the therm struct, decompation and backstripping outputs in the form of subsidence_strat_depths, Sμ, and Sσ, and prior estimates for stretching factor Beta (beta_ip) and time of thermal subsedence onset (t0_ip).

Examples:

(subsmdl, agedist, lldist, beta_t0dist, lldist_burnin) = SubsidenceStratMetropolis(smpl, config, therm, subsidence_strat_depths, Sμ, Sσ_corr, Beta_sigma/10, T0_sigma/10)

SubsidenceChron.find_formation_depths — Method

find_formation_depths(formation, thickness)

Find the top and bottom depth of each unique formation given a list of stratigraphic units and thicknesses

Examples:

ds = importdataset(joinpath("..", "examples", "Svalbard_highres.csv"), importas=:Tuple)
formations, tops bottoms = find_formation_depths(ds.Formation, ds.Thickness)

SubsidenceChron.find_formation_heights — Method

find_formation_heights(formation, thickness)

Find the top and bottom height of each unique formation given a list of stratigraphic units and thicknesses

Examples:

ds = importdataset(joinpath("..", "examples", "Svalbard_highres.csv"), importas=:Tuple)
formations, tops bottoms = find_formation_heights(ds.Formation, ds.Thickness)

SubsidenceChron.subsidenceparams — Method

subsidenceparams(lithology_string::AbstractString)

Return a tuple of the porosity-depth coefficient [1/m], surface porosity, and solid density of the given lithology. Supported lithologies include:

shale
black shale
siltstone
shaly sandstone
sandstone
chalk
limestone
dolostone
anhydrite
quartzite
diamictite
conglomerate
breccia
diabase
basalt
andesite
rhyolite
tuff